Melhorando a transicao de uma cena para outra pois a mesma esta muito travada e problematica. Talvez mais tarde seja viavel mudar bem o codigo da gamescene para deixa - la mais enxuta.

+++ Teste_sentaII/SentaHelper.swift
@@ -25,7 +25,7 @@ enum SentatrainStage: Int {
 class SentaHelper {
 
     static var tamanhoBarraDeProgresso: Float = 0
-    static var porcentagemGeralDoPrgoresso: Float = 0
+    static var porcentagemGeralDoPrgoresso: CGFloat = 0
     
     
     //Armazena o estagio de treinamento do cachorro
